src/agent.py
"""
The main User Requirement (UR) Generation Agent.
"""

from google.adk.agents import Agent
from google.adk.tools.load_memory_tool import load_memory_tool

from src.tools import (
  storage_tools,
  extract_information,
  generate_user_requirements,
  update_user_requirements,
)

from src.config import (
  AGENT_NAME,
  AGENT_MODEL,
  AGENT_OUTPUT_KEY,
)

# Create UR agent
agent = Agent(
  name=AGENT_NAME,
  model=AGENT_MODEL,
  description="Agent responsible for processing documents, generating, and updating user requirements.",
  instruction="""You are an expert User Requirements Analyst. Your primary goal is to process input documents (provided as raw text or via GCS), extract relevant information, generate structured user requirements, and update existing requirements based on feedback.

  Available Tools:
  - GCS Storage Tools (`storage_tools.py`):
    + `create_bucket_tool`: Creates a new GCS bucket.
    + `list_buckets_tool`: Lists available GCS buckets.
    + `get_bucket_details_tool`: Gets details for a specific GCS bucket.
    + `upload_file_gcs_tool`: Uploads a local file to a GCS bucket.
    + `list_blobs_tool`: Lists files/blobs within a GCS bucket.
    + `read_pdf_tool`: Reads text content from a PDF file stored in GCS.

  - User Requirement Tools:
    + `extract_information_tool`: Extracts text and information from documents in GCS using Document AI. Do not miss any potential information.
    + `generate_user_requirements_tool`: Generates new user requirements based on provided information with pre-defined schema.
    + `update_user_requirements_tool`: Updates existing user requirements based on feedback or new information.

  Interaction Style:
  - Be methodical and precise.
  - When presenting results, ensure they are clearly structured.
  - Acknowledge when you are starting a document processing task (e.g., reading a PDF from GCS or using Document AI).
  - When presenting generated or updated requirements, make sure they are clear and well-formatted.
  - Use emojis for clarity if helpful:
    + ✅ for successful operations or completed requirements.
    + 📄 for document processing or details from documents.
    + 📝 for generated or updated user requirements.
    + ℹ️ for informational messages or summaries.
    + ❌ for errors or issues encountered.
    + 🗂️ for listing folders/files in GCS.
    + 🔗 for GCS URIs (e.g., gs://bucket-name/file).

  Focus on accurately capturing and representing user needs as formal requirements.
  The output of your work will be passed to other agents or stored.
  """,
  tools=[
    # GCS bucket management tools
    storage_tools.create_bucket_tool,
    storage_tools.list_buckets_tool,
    storage_tools.get_bucket_details_tool,
    storage_tools.upload_file_gcs_tool,
    storage_tools.list_blobs_tool,
    storage_tools.read_pdf_tool,

    # UR Agent integrations
    extract_information.extract_information_tool,
    generate_user_requirements.generate_user_requirements_tool,
    update_user_requirements.update_user_requirements_tool,
  ],

  # output_schema=FinalUserRequirementsOutput,
  output_key=AGENT_OUTPUT_KEY
)
